refactor: move persistence wiring into services
